Moved equatable comformance into extension folder.
