+++ Sources/SunKit/Sun/Extensions/Sun+Equatable.swift
@@ -1,0 +1,28 @@
+//
+//  Sun+Equatable.swift
+//
+//
+//   Copyright 2024 Leonardo Bertinelli, Davide Biancardi, Raffaele Fulgente, Clelia Iovine, Nicolas Mariniello, Fabio Pizzano
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+import Foundation
+
+
+extension Sun: Equatable {
+    public static func == (lhs: Sun, rhs: Sun) -> Bool {
+        lhs.location == rhs.location &&
+        lhs.timeZone == rhs.timeZone &&
+        lhs.date == rhs.date
+    }
+}
